refactor: pubsub and bigquery use common build info
