+++ google/cloud/bigquery/bigquery_client.bzl
@@ -20,8 +20,8 @@ bigquery_client_hdrs = [
     "client.h",
     "connection.h",
     "connection_options.h",
-    "internal/storage_stub.h",
     "internal/connection_impl.h",
+    "internal/storage_stub.h",
     "internal/stream_reader.h",
     "internal/streaming_read_result_source.h",
     "read_result.h",
@@ -35,8 +35,8 @@ bigquery_client_hdrs = [
 bigquery_client_srcs = [
     "client.cc",
     "connection_options.cc",
-    "internal/storage_stub.cc",
     "internal/connection_impl.cc",
+    "internal/storage_stub.cc",
     "internal/streaming_read_result_source.cc",
     "read_stream.cc",
     "version.cc",
