--- google/cloud/pubsub/internal/build_info.cc.in
@@ -1,44 +1,0 @@
-// Copyright 2020 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include "google/cloud/pubsub/internal/build_info.h"
-#include "google/cloud/internal/port_platform.h"
-#include <algorithm>
-#include <cctype>
-#include <iterator>
-
-namespace google {
-namespace cloud {
-namespace pubsub_internal {
-inline namespace GOOGLE_CLOUD_CPP_PUBSUB_NS {
-
-std::string BuildFlags() {
-  static auto const* const kFlags = [] {
-    auto* flags = new std::string(R"""(@CMAKE_CXX_FLAGS@)""");
-    if (!flags->empty()) *flags += ' ';
-    *flags +=
-        R"""(${CMAKE_CXX_FLAGS_${GOOGLE_CLOUD_CPP_PUBSUB_BUILD_TYPE_UPPER}})""";
-    return flags;
-  }();
-  return *kFlags;
-}
-
-std::string BuildMetadata() {
-  return R"""(@GOOGLE_CLOUD_CPP_PUBSUB_BUILD_METADATA@)""";
-}
-
-}  // namespace GOOGLE_CLOUD_CPP_PUBSUB_NS
-}  // namespace pubsub_internal
-}  // namespace cloud
-}  // namespace google
