--- google/cloud/pubsub/internal/build_info_test.cc
@@ -1,42 +1,0 @@
-// Copyright 2020 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include "google/cloud/pubsub/internal/build_info.h"
-#include <gmock/gmock.h>
-
-namespace google {
-namespace cloud {
-namespace pubsub_internal {
-inline namespace GOOGLE_CLOUD_CPP_PUBSUB_NS {
-namespace {
-
-using ::testing::HasSubstr;
-using ::testing::Not;
-
-TEST(BuildInfo, BuildFlags) {
-  auto bf = BuildFlags();
-  EXPECT_THAT(bf, Not(HasSubstr("@")));
-}
-
-TEST(BuildInfo, BuildMetadata) {
-  auto const md = BuildMetadata();
-  EXPECT_FALSE(md.empty());
-  EXPECT_THAT(md, Not(HasSubstr("@")));
-}
-
-}  // namespace
-}  // namespace GOOGLE_CLOUD_CPP_PUBSUB_NS
-}  // namespace pubsub_internal
-}  // namespace cloud
-}  // namespace google
