+++ google/cloud/pubsub/pubsub_client.bzl
@@ -21,7 +21,6 @@ pubsub_client_hdrs = [
     "create_subscription_builder.h",
     "create_topic_builder.h",
     "internal/build_info.h",
-    "internal/compiler_info.h",
     "internal/publisher_stub.h",
     "internal/subscriber_stub.h",
     "internal/user_agent_prefix.h",
@@ -37,7 +36,6 @@ pubsub_client_hdrs = [
 
 pubsub_client_srcs = [
     "connection_options.cc",
-    "internal/compiler_info.cc",
     "internal/publisher_stub.cc",
     "internal/subscriber_stub.cc",
     "internal/user_agent_prefix.cc",
