+++ google/cloud/pubsub/pubsub_client_unit_tests.bzl
@@ -19,8 +19,6 @@
 pubsub_client_unit_tests = [
     "create_subscription_builder_test.cc",
     "create_topic_builder_test.cc",
-    "internal/build_info_test.cc",
-    "internal/compiler_info_test.cc",
     "internal/user_agent_prefix_test.cc",
     "subscription_test.cc",
     "topic_test.cc",
